Complete rename of repository
Rename classes and update POM file for transfer to SPDX repo
