--- pom-orig.xml
@@ -1,208 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <groupId>com.sourceauditor</groupId>
-  <artifactId>spdxcyclone</artifactId>
-  <version>0.0.1-SNAPSHOT</version>
-
-  <name>spdxcyclone</name>
-  <url>https://github.com/goneall/spdxcyclone</url>
-
-  <properties>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-  </properties>
-
-  <dependencies>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <version>4.11</version>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-    	<groupId>org.spdx</groupId>
-    	<artifactId>java-spdx-library</artifactId>
-    	<version>1.0.10</version>
-    </dependency>
-    <dependency>
-    	<groupId>org.cyclonedx</groupId>
-    	<artifactId>cyclonedx-core-java</artifactId>
-    	<version>5.0.2</version>
-    </dependency>
-    <dependency>
-    	<groupId>org.spdx</groupId>
-    	<artifactId>spdx-rdf-store</artifactId>
-    	<version>1.0.4</version>
-    </dependency>
-    <dependency>
-    	<groupId>org.spdx</groupId>
-    	<artifactId>spdx-jackson-store</artifactId>
-    	<version>1.0.4</version>
-    </dependency>
-    <dependency>
-    	<groupId>org.spdx</groupId>
-    	<artifactId>spdx-tagvalue-store</artifactId>
-    	<version>1.0.4</version>
-    </dependency>
-    <dependency>
-    	<groupId>org.spdx</groupId>
-    	<artifactId>spdx-spreadsheet-store</artifactId>
-    	<version>1.0.3</version>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <testResources>
-      <testResource>
-		<directory>src/test/resources</directory>
-	</testResource>
-    </testResources>
-    <pluginManagement><!-- lock down plugins versions to avoid using Maven defaults (may be moved to parent pom) -->
-      <plugins>
-        <!-- clean lifecycle, see https://maven.apache.org/ref/current/maven-core/lifecycles.html#clean_Lifecycle -->
-        <plugin>
-          <artifactId>maven-clean-plugin</artifactId>
-          <version>3.1.0</version>
-        </plugin>
-        <!-- default lifecycle, jar packaging: see https://maven.apache.org/ref/current/maven-core/default-bindings.html#Plugin_bindings_for_jar_packaging -->
-        <plugin>
-          <artifactId>maven-resources-plugin</artifactId>
-          <version>3.0.2</version>
-        </plugin>
-        <plugin>
-          <artifactId>maven-compiler-plugin</artifactId>
-          <version>3.8.0</version>
-          <configuration>
-            <release>11</release>
-          </configuration>
-        </plugin>
-        <plugin>
-		  <artifactId>maven-assembly-plugin</artifactId>
-		  <configuration>
-		    <archive>
-		      <manifest>
-		        <mainClass>com.sourceauditor.spdxcyclone.Main</mainClass>
-		      </manifest>
-		    </archive>
-		    <descriptorRefs>
-		      <descriptorRef>jar-with-dependencies</descriptorRef>
-		    </descriptorRefs>
-		  </configuration>
-		  <executions>
-		    <execution>
-		      <id>make-assembly</id>
-		      <phase>package</phase>
-		      <goals>
-		        <goal>single</goal>
-		      </goals>
-		    </execution>
-		  </executions>
-		</plugin>
-		<plugin>
-			<groupId>org.apache.maven.plugins</groupId>
-			<artifactId>maven-javadoc-plugin</artifactId>
-			<version>2.9</version>
-			<configuration>
-				<quiet>true</quiet>
-				<source>8</source>
-				<javadocExecutable>${env.JAVA_HOME}/bin/javadoc</javadocExecutable>
-				<additionalparam>-Xdoclint:none</additionalparam>
-			</configuration>
-			<executions>
-				<execution>
-					<id>attach-javadocs</id>
-					<configuration>
-						<additionalparam>${javadoc.opts}</additionalparam>
-					</configuration>
-					<goals>
-						<goal>jar</goal>
-					</goals>
-				</execution>
-			</executions>
-		</plugin>
-	   	<plugin>
-	      <groupId>org.apache.maven.plugins</groupId>
-	      <artifactId>maven-source-plugin</artifactId>
-	      <version>3.2.1</version>
-	      <executions>
-	        <execution>
-	          <id>attach-sources</id>
-	          <phase>verify</phase>
-	          <goals>
-	            <goal>jar-no-fork</goal>
-	          </goals>
-	        </execution>
-	      </executions>
-	    </plugin>
-        <plugin>
-          <artifactId>maven-surefire-plugin</artifactId>
-          <version>2.22.1</version>
-        </plugin>
-        <plugin>
-          <artifactId>maven-jar-plugin</artifactId>
-          <version>3.0.2</version>
-        </plugin>
-        <plugin>
-          <artifactId>maven-install-plugin</artifactId>
-          <version>2.5.2</version>
-        </plugin>
-        <plugin>
-          <artifactId>maven-deploy-plugin</artifactId>
-          <version>2.8.2</version>
-        </plugin>
-        <!-- site lifecycle, see https://maven.apache.org/ref/current/maven-core/lifecycles.html#site_Lifecycle -->
-        <plugin>
-          <artifactId>maven-site-plugin</artifactId>
-          <version>3.7.1</version>
-        </plugin>
-        <plugin>
-          <artifactId>maven-project-info-reports-plugin</artifactId>
-          <version>3.0.0</version>
-        </plugin>
-		<plugin>
-			<groupId>org.spdx</groupId>
-				<artifactId>spdx-maven-plugin</artifactId>
-				<version>0.5.5</version>
-				<executions>
-					<execution>
-						<id>build-spdx</id>
-						<phase>prepare-package</phase>
-						<goals>
-							<goal>createSPDX</goal>
-						</goals>
-					</execution>
-				</executions>
-				<configuration>
-					<spdxDocumentNamespace>http://sourceauditor.com/spdxdocs/spdxcyclonedx-{$version}</spdxDocumentNamespace>					
-					<defaultFileCopyright>Copyright (c) 2022 Source Auditor Inc.</defaultFileCopyright>
-					<defaultFileContributors>
-						<param>Gary O'Neall</param>
-					</defaultFileContributors>
-					<defaultLicenseInformationInFile>Apache-2.0</defaultLicenseInformationInFile>
-					<defaultFileConcludedLicense>Apache-2.0</defaultFileConcludedLicense>
-					<defaultFileNotice>Licensed under the Apache License, Version 2.0 (the "License");
-	  you may not use this file except in compliance with the License.
-	  You may obtain a copy of the License at
-	
-	      http://www.apache.org/licenses/LICENSE-2.0
-	
-	  Unless required by applicable law or agreed to in writing, software
-	  distributed under the License is distributed on an "AS IS" BASIS,
-	  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-	  See the License for the specific language governing permissions and
-	  limitations under the License.</defaultFileNotice>
-					<licenseDeclared>Apache-2.0</licenseDeclared>
-					<licenseConcluded>Apache-2.0</licenseConcluded>
-					<creators>
-						<param>Person: Gary O'Neall</param>
-					</creators>
-					<originator>Organization: Linux Foundation</originator>
-				</configuration>
-          </plugin>
-      </plugins>
-    </pluginManagement>
-  </build>
-</project>
