+++ src/main/java/org/spdx/cdx2spdx/CycloneConversionException.java
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package com.sourceauditor.spdxcyclone;
+package org.spdx.cdx2spdx;
 
 /**
  * Exceptions related to incompatible or missing properties when converting from CycloneDx to SPDX
